imuの動作確認完了

+++ scripts/after.py
@@ -2,4 +2,5 @@
 import os
 
 os.chdir(os.path.dirname(os.path.abspath(__file__)))
-os.rename("./../Core/Src/main.c", "./../Core/Src/main.cpp")
+if os.path.exists("./../Core/Src/main.c"):
+    os.rename("./../Core/Src/main.c", "./../Core/Src/main.cpp")
